Add files via upload
Our Basic Calculator is the perfect tool for anyone looking to perform basic arithmetic calculations quickly and easily. With a large and easy-to-use interface, this calculator is simple and intuitive, making it perfect for users of all ages and levels of experience.

The best thing about our Basic Calculator? It is completely free to use online, so you can perform calculations anytime and anywhere without the need for additional software or hardware. Whether you're a student studying math or a professional in a field that requires basic calculations, our Basic Calculator is a valuable resource that can help make your work easier and more efficient.

This calculator has only the basic mathematics functions...